+++ src/test/OpenEXRTest/testPreviewImage.cpp
@@ -119,7 +119,7 @@ readWriteFiles (
         file2.setFrameBuffer (pixels2 - dx - dy * w, 1, w);
         file2.readPixels (dw.min.y, dw.max.y);
 
-        for (size_t i = 0; i < w * h; ++i)
+        for (int i = 0; i < w * h; ++i)
         {
             assert (pixels1[i].r == pixels2[i].r);
             assert (pixels1[i].g == pixels2[i].g);
